Moved code source to UI folder and out of UI projekt-monsterz-front.
