+++ ui/package.json
@@ -18,6 +18,8 @@
     "@angular/platform-browser": "^15.1.0",
     "@angular/platform-browser-dynamic": "^15.1.0",
     "@angular/router": "^15.1.0",
+    "@gsa-sam/ngx-uswds": "^13.0.0-beta.2",
+    "@trussworks/react-uswds": "^4.0.0",
     "@uswds/uswds": "^3.3.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
