+++ ui/src/app/app.module.ts
@@ -9,7 +9,6 @@ import { WelcomePageComponent } from './welcome-page/welcome-page.component';
 import { FormsModule } from '@angular/forms';
 import { FooterComponent } from './footer/footer.component';
 
-
 @NgModule({
   declarations: [
     AppComponent,
