+++ ui/src/app/footer/footer.component.spec.ts
