+++ ui/src/app/footer/footer.component.ts
