+++ ui/src/app/welcome-page/welcome-page.component.spec.ts
