+++ ui/src/app/welcome-page/welcome-page.component.ts
