+++ ui/src/assets/welcome-screen-monsters/monster0_mouth.png
